mcp: move example agent tools into importable module
